Refactor: Replace bundle with ES modules
- Remove app.bundle.js and package-lock.json
- Add new modules: camera.js, controls.js, input.js, simulation.js
- Update main.js, renderer.js, index.html, package.json
