+++ index.html
@@ -495,7 +495,7 @@
       <button id="btnZoomIn" class="wide" title="Zoom in">+</button>
     </div>
 
-    <script src="./app.bundle.js"></script>
+    <script type="module" src="./src/main.js"></script>
     <script>
       // Panel toggles (independent; each can occupy full height)
       const overlayControls = document.getElementById("overlayControls");
